# Support more types for normals