This involved removing numpy specific typing from the Cython
and thus making the code more 'c'-like. However, this had an
advantage of making the code faster. While I was here I also
seperated the vertex and face calculation to make face normal
calculation from Python twice as fast.

This fixes #718.

The speedup is as follows:
```
OLD
100 loops, best of 3: 10.2 ms per loop
100 loops, best of 3: 9.53 ms per loop
```
```
NEW
100 loops, best of 3: 8.28 ms per loop
100 loops, best of 3: 4.93 ms per loop
```