@patricksnape 2016-08-01
We can make this a little bit faster by turning off bounds checking - though at the risk of a segfault if the triangulation is bad.