@patricksnape 2016-08-05
@jabooth Please review.